+++ tests/unit/project_test.py
@@ -10,8 +10,8 @@ from .. import unittest
 from compose.common.const import LABEL_SERVICE
 from compose.config.config import Config
 from compose.config.types import VolumeFromSpec
-from compose.container import Container
 from compose.core import dockerclient as dc
+from compose.core.container import Container
 from compose.project import Project
 from compose.service import ImageType
 from compose.service import Service
